The initial commit of this file had minimal ports assigned to the module. Upon designing the BlazeRouter flow diagram, I found additional ports that were required for operation with the rest of the circuit as a whole.
Most all of the modules (.vhd) file do not have any logic tied to it.
In this commit, I did add some code to the LinkController to verify some errors
displayed during mapping and implementation. Therefore the majority of the modules are black box
implementation and will not pass synthesis. You must add code in the
architecture portion of each .vhd module, simulate the RTL and then
synthesize to continue.